chore: Split crate into separate bin and lib targets
Splits the crate into separate lib and bin targets to set the stage for
adding integration tests.
